Fix: rename command
