+++ src/commands/index.ts
@@ -11,7 +11,7 @@ import { generateUUID } from "./generateUUID";
 import { generateHandle } from './generateHandle';
 import { convertVersionNumber } from './convertVersionNumber';
 import { bumpVersionNumber } from './bumpVersionNumber';
-import { initializeLuaConfiguration } from './luaConfiguration';
+import { initializeLuaConfiguration } from './initializeLuaConfiguration';
 
 /** Registers the command to VS Code */
 function registerCommand(command: (...args: any[]) => any): vscode.Disposable {
